Reorganizing file structure, package names and import statements to new package name.
