+++ example/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.slidingmenu.example"
+    package="com.jeremyfeinstein.slidingmenu.example"
     android:versionCode="3"
     android:versionName="1.2" >
