+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/AttachExample.java
