+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/BaseActivity.java
