+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/LeftAndRightActivity.java
