+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/PropertiesActivity.java
