+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/SlidingContent.java
