+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/SlidingTitleBar.java
