+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/Util.java
