+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/ViewPagerActivity.java
