+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/anim/CustomScaleAnimation.java
