+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/anim/CustomSlideAnimation.java
