+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/anim/CustomZoomAnimation.java
