+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/fragments/BirdActivity.java
