+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/fragments/BirdMenuFragment.java
