+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/fragments/ColorMenuFragment.java
