+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/fragments/FragmentChangeActivity.java
