+++ example/src/main/java/com/jeremyfeinstein/slidingmenu/example/fragments/ResponsiveUIActivity.java
