+++ library/src/main/java/com/jeremyfeinstein/slidingmenu/lib/app/SlidingPreferenceActivity.java
@@ -1,15 +1,14 @@
-package com.slidingmenu.lib.app;
+package com.jeremyfeinstein.lib.app;
 
-import android.app.ListActivity;
+import com.jeremyfeinstein.lib.SlidingMenu;
+
 import android.os.Bundle;
+import android.preference.PreferenceActivity;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewGroup.LayoutParams;
-import android.widget.ListView;
 
-import com.slidingmenu.lib.SlidingMenu;
-
-public class SlidingListActivity extends ListActivity implements SlidingActivityBase {
+public class SlidingPreferenceActivity extends PreferenceActivity implements SlidingActivityBase {
 
 	private SlidingActivityHelper mHelper;
 
@@ -18,12 +17,9 @@ public class SlidingListActivity extends ListActivity implements SlidingActivity
 	 */
 	@Override
 	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
 		mHelper = new SlidingActivityHelper(this);
+		super.onCreate(savedInstanceState);
 		mHelper.onCreate(savedInstanceState);
-		ListView listView = new ListView(this);
-		listView.setId(android.R.id.list);
-		setContentView(listView);
 	}
 
 	/* (non-Javadoc)
@@ -81,64 +77,63 @@ public class SlidingListActivity extends ListActivity implements SlidingActivity
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#setBehindContentView(int)
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#setBehindContentView(int)
 	 */
 	public void setBehindContentView(int id) {
 		setBehindContentView(getLayoutInflater().inflate(id, null));
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#setBehindContentView(android.view.View)
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#setBehindContentView(android.view.View)
 	 */
 	public void setBehindContentView(View v) {
 		setBehindContentView(v, new LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT));
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#setBehindContentView(android.view.View, android.view.ViewGroup.LayoutParams)
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#setBehindContentView(android.view.View, android.view.ViewGroup.LayoutParams)
 	 */
 	public void setBehindContentView(View v, LayoutParams params) {
 		mHelper.setBehindContentView(v, params);
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#getSlidingMenu()
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#getSlidingMenu()
 	 */
 	public SlidingMenu getSlidingMenu() {
 		return mHelper.getSlidingMenu();
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#toggle()
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#toggle()
 	 */
 	public void toggle() {
 		mHelper.toggle();
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#showAbove()
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#showAbove()
 	 */
 	public void showContent() {
 		mHelper.showContent();
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#showBehind()
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#showBehind()
 	 */
 	public void showMenu() {
 		mHelper.showMenu();
 	}
 	
-	/*
-	 * (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#showSecondaryMenu()
+	/* (non-Javadoc)
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#showSecondaryMenu()
 	 */
 	public void showSecondaryMenu() {
 		mHelper.showSecondaryMenu();
 	}
 
 	/* (non-Javadoc)
-	 * @see com.slidingmenu.lib.app.SlidingActivityBase#setSlidingActionBarEnabled(boolean)
+	 * @see com.jeremyfeinstein.lib.app.SlidingActivityBase#setSlidingActionBarEnabled(boolean)
 	 */
 	public void setSlidingActionBarEnabled(boolean b) {
 		mHelper.setSlidingActionBarEnabled(b);
@@ -153,5 +148,4 @@ public class SlidingListActivity extends ListActivity implements SlidingActivity
 		if (b) return b;
 		return super.onKeyUp(keyCode, event);
 	}
-
 }
