+++ pom.xml
@@ -3,7 +3,7 @@
 	         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.slidingmenu</groupId>
+	<groupId>com.jeremyfeinstein.slidingmenu</groupId>
 	<artifactId>parent</artifactId>
 	<version>1.3-SNAPSHOT</version>
 	<packaging>pom</packaging>
